refactor: Moving src/api.js into src/api/gemini.js
